+++ src/features/vehicles/components/VehicleTable.tsx
@@ -1,0 +1,339 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import { Eye, Edit, Plus, ChevronUp, ChevronDown, ChevronLeft, ChevronRight } from 'lucide-react';
+import { Vehicle, User } from '../../../types';
+import { Badge } from '../../../components/ui/Badge';
+import { LoadingSpinner } from '../../../components/ui/LoadingSpinner';
+import { Button } from '../../../components/ui/Button';
+
+interface VehicleTableProps {
+  vehicles: Vehicle[];
+  totalCount: number;
+  totalPages: number;
+  currentPage: number;
+  hasNextPage: boolean;
+  hasPreviousPage: boolean;
+  sortBy: string;
+  sortOrder: 'asc' | 'desc';
+  itemsPerPage: number;
+  loading: boolean;
+  user: User | null;
+  searchTerm: string;
+  statusFilters: string[];
+  unassignedFilter: boolean;
+  onSorting: (column: 'name' | 'status' | 'mileage' | 'maintenanceCost' | 'assignedDriver') => void;
+  onCurrentPageChange: (page: number) => void;
+  onItemsPerPageChange: (limit: number) => void;
+  onClearAllFilters: () => void;
+}
+
+export function VehicleTable({
+  vehicles,
+  totalCount,
+  totalPages,
+  currentPage,
+  hasNextPage,
+  hasPreviousPage,
+  sortBy,
+  sortOrder,
+  itemsPerPage,
+  loading,
+  user,
+  searchTerm,
+  statusFilters,
+  unassignedFilter,
+  onSorting,
+  onCurrentPageChange,
+  onItemsPerPageChange,
+  onClearAllFilters,
+}: VehicleTableProps) {
+  // Render sort icon
+  const renderSortIcon = (column: string) => {
+    if (sortBy !== column) {
+      return <ChevronUp className="h-4 w-4 text-gray-300" />;
+    }
+    return sortOrder === 'asc' 
+      ? <ChevronUp className="h-4 w-4 text-blue-600" />
+      : <ChevronDown className="h-4 w-4 text-blue-600" />;
+  };
+
+  // Handle page change
+  const handlePageChange = (page: number) => {
+    if (page >= 1 && page <= totalPages) {
+      onCurrentPageChange(page);
+    }
+  };
+
+  // Generate page numbers for pagination
+  const getPageNumbers = () => {
+    const pages = [];
+    const maxVisiblePages = 5;
+    
+    if (totalPages <= maxVisiblePages) {
+      for (let i = 1; i <= totalPages; i++) {
+        pages.push(i);
+      }
+    } else {
+      const startPage = Math.max(1, currentPage - Math.floor(maxVisiblePages / 2));
+      const endPage = Math.min(totalPages, startPage + maxVisiblePages - 1);
+      
+      for (let i = startPage; i <= endPage; i++) {
+        pages.push(i);
+      }
+    }
+    
+    return pages;
+  };
+
+  return (
+    <div className="bg-white shadow rounded-lg">
+      <div className="px-4 py-3 sm:px-6 border-b border-gray-200">
+        <div className="flex items-center justify-between">
+          <div className="flex items-center space-x-4">
+            <h3 className="text-lg font-medium text-gray-900">
+              Fleet Vehicles ({totalCount})
+            </h3>
+            {loading && <LoadingSpinner size="sm" />}
+          </div>
+          
+          <div className="flex items-center space-x-4">
+            {/* Items per page */}
+            <div className="flex items-center space-x-2">
+              <label htmlFor="itemsPerPage" className="text-sm text-gray-700">
+                Show:
+              </label>
+              <select
+                id="itemsPerPage"
+                value={itemsPerPage}
+                onChange={(e) => onItemsPerPageChange(Number(e.target.value))}
+                className="border border-gray-300 rounded-md text-sm py-1 px-2 focus:outline-none focus:ring-1 focus:ring-blue-500 focus:border-blue-500"
+              >
+                <option value={5}>5</option>
+                <option value={10}>10</option>
+                <option value={25}>25</option>
+                <option value={50}>50</option>
+              </select>
+            </div>
+            
+            {totalCount > 0 && (
+              <div className="text-sm text-gray-500">
+                Showing {((currentPage - 1) * itemsPerPage) + 1}-{Math.min(currentPage * itemsPerPage, totalCount)} of {totalCount}
+              </div>
+            )}
+          </div>
+        </div>
+      </div>
+
+      {/* Vehicles Table */}
+      <div className="overflow-x-auto">
+        {vehicles.length > 0 ? (
+          <table className="min-w-full divide-y divide-gray-200">
+            <thead className="bg-gray-50">
+              <tr>
+                <th 
+                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
+                  onClick={() => onSorting('name')}
+                >
+                  <div className="flex items-center space-x-1">
+                    <span>Vehicle Name</span>
+                    {renderSortIcon('name')}
+                  </div>
+                </th>
+                <th 
+                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
+                  onClick={() => onSorting('status')}
+                >
+                  <div className="flex items-center space-x-1">
+                    <span>Status</span>
+                    {renderSortIcon('status')}
+                  </div>
+                </th>
+                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                  VIN
+                </th>
+                <th 
+                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
+                  onClick={() => onSorting('mileage')}
+                >
+                  <div className="flex items-center space-x-1">
+                    <span>Mileage</span>
+                    {renderSortIcon('mileage')}
+                  </div>
+                </th>
+                <th 
+                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
+                  onClick={() => onSorting('maintenanceCost')}
+                >
+                  <div className="flex items-center space-x-1">
+                    <span>Maintenance Cost</span>
+                    {renderSortIcon('maintenanceCost')}
+                  </div>
+                </th>
+                <th 
+                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
+                  onClick={() => onSorting('assignedDriver')}
+                >
+                  <div className="flex items-center space-x-1">
+                    <span>Assigned Driver</span>
+                    {renderSortIcon('assignedDriver')}
+                  </div>
+                </th>
+                <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
+                  Actions
+                </th>
+              </tr>
+            </thead>
+            <tbody className="bg-white divide-y divide-gray-200">
+              {vehicles.map((vehicle) => (
+                <tr key={vehicle.id} className="hover:bg-gray-50 transition-colors duration-150">
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div>
+                      <div className="text-sm font-medium text-gray-900">
+                        {vehicle.name}
+                      </div>
+                      <div className="text-sm text-gray-500">
+                        {vehicle.make} {vehicle.model} {vehicle.year}
+                      </div>
+                    </div>
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <Badge 
+                      type={vehicle.status === 'active' ? 'green' : vehicle.status === 'maintenance' ? 'orange' : 'red'} 
+                      label={vehicle.status === 'active' ? 'Active' : vehicle.status === 'maintenance' ? 'Maintenance' : 'Idle'} 
+                    />
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900 font-mono">
+                    {vehicle.vin || 'N/A'}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
+                    {vehicle.mileage?.toLocaleString() || 'N/A'} miles
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
+                    ${vehicle.maintenanceCost?.toLocaleString() || '0'}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    {vehicle.assignedDriverName ? (
+                      <div>
+                        <div className="text-sm font-medium text-gray-900">
+                          {vehicle.assignedDriverName}
+                        </div>
+                        <div className="text-sm text-gray-500">
+                          {vehicle.assignedDriverEmail || 'No email'}
+                        </div>
+                      </div>
+                    ) : (
+                      <Badge type="gray" label="Unassigned" />
+                    )}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                    <div className="flex justify-end space-x-2">
+                      <Link
+                        to={`/vehicles/${vehicle.id}`}
+                        className="text-blue-600 hover:text-blue-900 p-1 hover:bg-blue-50 rounded transition-colors duration-200 focus-ring"
+                        title="View Details"
+                      >
+                        <Eye className="h-4 w-4" />
+                      </Link>
+                      {user?.isAdmin && (
+                        <Link
+                          to={`/vehicles/${vehicle.id}/edit`}
+                          className="text-gray-600 hover:text-gray-900 p-1 hover:bg-gray-50 rounded transition-colors duration-200 focus-ring"
+                          title="Edit Vehicle"
+                        >
+                          <Edit className="h-4 w-4" />
+                        </Link>
+                      )}
+                    </div>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        ) : (
+          <div className="text-center py-12">
+            <div className="text-gray-500 mb-4">
+              {loading ? 'Loading vehicles...' : 
+               (searchTerm || statusFilters.length > 0 || unassignedFilter) ? 
+               'No vehicles match your filters' : 'No vehicles found'}
+            </div>
+            {!loading && (searchTerm || statusFilters.length > 0 || unassignedFilter) && (
+              <Button
+                onClick={onClearAllFilters}
+                variant="secondary"
+              >
+                Clear all filters
+              </Button>
+            )}
+          </div>
+        )}
+      </div>
+
+      {/* Pagination */}
+      {totalPages > 1 && (
+        <div className="bg-white px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6">
+          <div className="flex-1 flex justify-between sm:hidden">
+            <button
+              onClick={() => handlePageChange(currentPage - 1)}
+              disabled={!hasPreviousPage || loading}
+              className="relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+            >
+              Previous
+            </button>
+            <button
+              onClick={() => handlePageChange(currentPage + 1)}
+              disabled={!hasNextPage || loading}
+              className="ml-3 relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+            >
+              Next
+            </button>
+          </div>
+          <div className="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
+            <div>
+              <p className="text-sm text-gray-700">
+                Showing <span className="font-medium">{((currentPage - 1) * itemsPerPage) + 1}</span> to{' '}
+                <span className="font-medium">{Math.min(currentPage * itemsPerPage, totalCount)}</span> of{' '}
+                <span className="font-medium">{totalCount}</span> results
+              </p>
+            </div>
+            <div>
+              <nav className="relative z-0 inline-flex rounded-md shadow-sm -space-x-px" aria-label="Pagination">
+                <button
+                  onClick={() => handlePageChange(currentPage - 1)}
+                  disabled={!hasPreviousPage || loading}
+                  className="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+                >
+                  <span className="sr-only">Previous</span>
+                  <ChevronLeft className="h-5 w-5" aria-hidden="true" />
+                </button>
+                
+                {getPageNumbers().map((pageNumber) => (
+                  <button
+                    key={pageNumber}
+                    onClick={() => handlePageChange(pageNumber)}
+                    disabled={loading}
+                    className={`relative inline-flex items-center px-4 py-2 border text-sm font-medium ${
+                      currentPage === pageNumber
+                        ? 'z-10 bg-blue-50 border-blue-500 text-blue-600'
+                        : 'bg-white border-gray-300 text-gray-500 hover:bg-gray-50'
+                    } disabled:opacity-50 disabled:cursor-not-allowed`}
+                  >
+                    {pageNumber}
+                  </button>
+                ))}
+                
+                <button
+                  onClick={() => handlePageChange(currentPage + 1)}
+                  disabled={!hasNextPage || loading}
+                  className="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+                >
+                  <span className="sr-only">Next</span>
+                  <ChevronRight className="h-5 w-5" aria-hidden="true" />
+                </button>
+              </nav>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
