+++ src/pages/Vehicles.tsx
@@ -1,513 +1,6 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import { Eye, Edit, Plus, Search, Filter, ChevronUp, ChevronDown, ChevronLeft, ChevronRight, X, RotateCcw } from 'lucide-react';
-import { useAuth } from '../hooks/useAuth';
-import { useVehiclesData } from '../hooks/useVehiclesData';
-import { Badge } from '../components/ui/Badge';
-import { LoadingSpinner } from '../components/ui/LoadingSpinner';
-import { Button } from '../components/ui/Button';
+import { VehiclesPage } from '../features/vehicles/pages/VehiclesPage';
 
 export function Vehicles() {
-  const { user } = useAuth();
-  const [showFilterModal, setShowFilterModal] = React.useState(false);
-  
-  const {
-    // Data
-    vehicles,
-    totalCount,
-    totalPages,
-    currentPage,
-    hasNextPage,
-    hasPreviousPage,
-    
-    // State
-    searchTerm,
-    statusFilters,
-    unassignedFilter,
-    sortBy,
-    sortOrder,
-    itemsPerPage,
-    loading,
-    error,
-    
-    // Actions
-    setSearchTerm,
-    toggleStatusFilter,
-    clearStatusFilters,
-    setUnassignedFilter,
-    setSorting,
-    setCurrentPage,
-    setItemsPerPage,
-    clearAllFilters,
-    refreshData,
-  } = useVehiclesData();
-
-  // Enhanced clear all filters function that also closes the modal
-  const handleClearAllFilters = () => {
-    clearAllFilters();
-    setShowFilterModal(false);
-  };
-
-  // Render sort icon
-  const renderSortIcon = (column: string) => {
-    if (sortBy !== column) {
-      return <ChevronUp className="h-4 w-4 text-gray-300" />;
-    }
-    return sortOrder === 'asc' 
-      ? <ChevronUp className="h-4 w-4 text-blue-600" />
-      : <ChevronDown className="h-4 w-4 text-blue-600" />;
-  };
-
-  // Handle page change
-  const handlePageChange = (page: number) => {
-    if (page >= 1 && page <= totalPages) {
-      setCurrentPage(page);
-    }
-  };
-
-  // Generate page numbers for pagination
-  const getPageNumbers = () => {
-    const pages = [];
-    const maxVisiblePages = 5;
-    
-    if (totalPages <= maxVisiblePages) {
-      for (let i = 1; i <= totalPages; i++) {
-        pages.push(i);
-      }
-    } else {
-      const startPage = Math.max(1, currentPage - Math.floor(maxVisiblePages / 2));
-      const endPage = Math.min(totalPages, startPage + maxVisiblePages - 1);
-      
-      for (let i = startPage; i <= endPage; i++) {
-        pages.push(i);
-      }
-    }
-    
-    return pages;
-  };
-
-  if (error) {
-    return (
-      <div className="space-y-6">
-        <div className="flex justify-between items-center">
-          <div>
-            <h1 className="text-2xl font-bold text-gray-900">Vehicles</h1>
-            <p className="mt-1 text-sm text-gray-600">Manage your fleet vehicles</p>
-          </div>
-        </div>
-        
-        <div className="bg-red-50 border border-red-200 rounded-lg p-4">
-          <div className="flex items-center justify-between">
-            <div className="text-red-800">
-              <h3 className="font-medium">Error loading vehicles</h3>
-              <p className="text-sm mt-1">{error}</p>
-            </div>
-            <button
-              onClick={refreshData}
-              className="text-red-600 hover:text-red-700 transition-colors duration-200"
-            >
-              <RotateCcw className="h-5 w-5" />
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6">
-      {/* Header */}
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-2xl font-bold text-gray-900">Vehicles</h1>
-          <p className="mt-1 text-sm text-gray-600">
-            Manage your fleet vehicles
-          </p>
-        </div>
-        <div className="flex items-center space-x-3">
-          {/* Filter Button */}
-          <Button
-            onClick={() => setShowFilterModal(true)}
-            variant="secondary"
-          >
-            <Filter className="h-4 w-4 mr-2" />
-            Filters
-            {(searchTerm || statusFilters.length > 0 || unassignedFilter) && (
-              <span className="ml-2 inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium bg-blue-100 text-blue-800">
-                Active
-              </span>
-            )}
-          </Button>
-          
-          {/* Add Vehicle Button */}
-          {user?.isAdmin && (
-            <Link
-              to="/vehicles/new"
-              className="btn-primary"
-            >
-              <Plus className="h-4 w-4 mr-2" />
-              Add Vehicle
-            </Link>
-          )}
-        </div>
-      </div>
-
-      {/* Filter Modal */}
-      {showFilterModal && (
-        <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full z-50 flex items-center justify-center">
-          <div className="relative bg-white rounded-lg shadow-xl max-w-4xl w-full mx-4 max-h-96 overflow-y-auto">
-            <div className="px-4 py-5 sm:p-6">
-              {/* Modal Header */}
-              <div className="flex items-center justify-between mb-6">
-                <div className="flex items-center">
-                  <Filter className="h-5 w-5 text-gray-400 mr-2" />
-                  <h3 className="text-lg font-medium text-gray-900">Search & Filter Vehicles</h3>
-                </div>
-                <button
-                  onClick={() => setShowFilterModal(false)}
-                  className="text-gray-400 hover:text-gray-600 transition-colors duration-200"
-                >
-                  <X className="h-6 w-6" />
-                </button>
-              </div>
-
-              {/* Filter Content */}
-              <div className="grid grid-cols-1 gap-4 lg:grid-cols-4">
-                {/* Search */}
-                <div className="lg:col-span-2">
-                  <label htmlFor="search" className="block text-sm font-medium text-gray-700 mb-1">
-                    Search Vehicles
-                  </label>
-                  <div className="relative">
-                    <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                      <Search className="h-4 w-4 text-gray-400" />
-                    </div>
-                    <input
-                      type="text"
-                      id="search"
-                      value={searchTerm}
-                      onChange={(e) => setSearchTerm(e.target.value)}
-                      className="block w-full pl-10 pr-3 py-2 border border-gray-300 rounded-md leading-5 bg-white placeholder-gray-500 focus:outline-none focus:placeholder-gray-400 focus:ring-1 focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-                      placeholder="Search by name, VIN, make, or model..."
-                    />
-                  </div>
-                </div>
-                
-                {/* Status Filters */}
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">
-                    Status Filter
-                  </label>
-                  <div className="space-y-2">
-                    {['active', 'maintenance', 'idle'].map((status) => (
-                      <label key={status} className="flex items-center">
-                        <input
-                          type="checkbox"
-                          checked={statusFilters.includes(status)}
-                          onChange={() => toggleStatusFilter(status)}
-                          className="h-4 w-4 text-blue-600 focus:ring-blue-500 border-gray-300 rounded"
-                        />
-                        <span className="ml-2 text-sm text-gray-700 capitalize">{status}</span>
-                      </label>
-                    ))}
-                  </div>
-                </div>
-                
-                {/* Additional Filters */}
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">
-                    Additional Filters
-                  </label>
-                  <div className="space-y-2">
-                    <label className="flex items-center">
-                      <input
-                        type="checkbox"
-                        checked={unassignedFilter}
-                        onChange={(e) => setUnassignedFilter(e.target.checked)}
-                        className="h-4 w-4 text-blue-600 focus:ring-blue-500 border-gray-300 rounded"
-                      />
-                      <span className="ml-2 text-sm text-gray-700">Unassigned only</span>
-                    </label>
-                  </div>
-                </div>
-              </div>
-
-              {/* Modal Actions */}
-              <div className="mt-6 flex items-center justify-between pt-4 border-t border-gray-200">
-                <div>
-                  {(searchTerm || statusFilters.length > 0 || unassignedFilter) && (
-                    <Button
-                      onClick={handleClearAllFilters}
-                      variant="secondary"
-                      className="px-3 py-2"
-                    >
-                      <X className="h-4 w-4 mr-2" />
-                      Clear All Filters
-                    </Button>
-                  )}
-                </div>
-                <Button
-                  onClick={() => setShowFilterModal(false)}
-                  variant="primary"
-                >
-                  Apply Filters
-                </Button>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Results Summary */}
-      <div className="bg-white shadow rounded-lg">
-        <div className="px-4 py-3 sm:px-6 border-b border-gray-200">
-          <div className="flex items-center justify-between">
-            <div className="flex items-center space-x-4">
-              <h3 className="text-lg font-medium text-gray-900">
-                Fleet Vehicles ({totalCount})
-              </h3>
-              {loading && <LoadingSpinner size="sm" />}
-            </div>
-            
-            <div className="flex items-center space-x-4">
-              {/* Items per page */}
-              <div className="flex items-center space-x-2">
-                <label htmlFor="itemsPerPage" className="text-sm text-gray-700">
-                  Show:
-                </label>
-                <select
-                  id="itemsPerPage"
-                  value={itemsPerPage}
-                  onChange={(e) => setItemsPerPage(Number(e.target.value))}
-                  className="border border-gray-300 rounded-md text-sm py-1 px-2 focus:outline-none focus:ring-1 focus:ring-blue-500 focus:border-blue-500"
-                >
-                  <option value={5}>5</option>
-                  <option value={10}>10</option>
-                  <option value={25}>25</option>
-                  <option value={50}>50</option>
-                </select>
-              </div>
-              
-              {totalCount > 0 && (
-                <div className="text-sm text-gray-500">
-                  Showing {((currentPage - 1) * itemsPerPage) + 1}-{Math.min(currentPage * itemsPerPage, totalCount)} of {totalCount}
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-
-        {/* Vehicles Table */}
-        <div className="overflow-x-auto">
-          {vehicles.length > 0 ? (
-            <table className="min-w-full divide-y divide-gray-200">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th 
-                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
-                    onClick={() => setSorting('name')}
-                  >
-                    <div className="flex items-center space-x-1">
-                      <span>Vehicle Name</span>
-                      {renderSortIcon('name')}
-                    </div>
-                  </th>
-                  <th 
-                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
-                    onClick={() => setSorting('status')}
-                  >
-                    <div className="flex items-center space-x-1">
-                      <span>Status</span>
-                      {renderSortIcon('status')}
-                    </div>
-                  </th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    VIN
-                  </th>
-                  <th 
-                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
-                    onClick={() => setSorting('mileage')}
-                  >
-                    <div className="flex items-center space-x-1">
-                      <span>Mileage</span>
-                      {renderSortIcon('mileage')}
-                    </div>
-                  </th>
-                  <th 
-                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
-                    onClick={() => setSorting('maintenanceCost')}
-                  >
-                    <div className="flex items-center space-x-1">
-                      <span>Maintenance Cost</span>
-                      {renderSortIcon('maintenanceCost')}
-                    </div>
-                  </th>
-                  <th 
-                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer hover:bg-gray-100 transition-colors duration-200"
-                    onClick={() => setSorting('assignedDriver')}
-                  >
-                    <div className="flex items-center space-x-1">
-                      <span>Assigned Driver</span>
-                      {renderSortIcon('assignedDriver')}
-                    </div>
-                  </th>
-                  <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Actions
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="bg-white divide-y divide-gray-200">
-                {vehicles.map((vehicle) => (
-                  <tr key={vehicle.id} className="hover:bg-gray-50 transition-colors duration-150">
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <div>
-                        <div className="text-sm font-medium text-gray-900">
-                          {vehicle.name}
-                        </div>
-                        <div className="text-sm text-gray-500">
-                          {vehicle.make} {vehicle.model} {vehicle.year}
-                        </div>
-                      </div>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <Badge 
-                        type={vehicle.status === 'active' ? 'green' : vehicle.status === 'maintenance' ? 'orange' : 'red'} 
-                        label={vehicle.status === 'active' ? 'Active' : vehicle.status === 'maintenance' ? 'Maintenance' : 'Idle'} 
-                      />
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900 font-mono">
-                      {vehicle.vin || 'N/A'}
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                      {vehicle.mileage?.toLocaleString() || 'N/A'} miles
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                      ${vehicle.maintenanceCost?.toLocaleString() || '0'}
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      {vehicle.assignedDriverName ? (
-                        <div>
-                          <div className="text-sm font-medium text-gray-900">
-                            {vehicle.assignedDriverName}
-                          </div>
-                          <div className="text-sm text-gray-500">
-                            {vehicle.assignedDriverEmail || 'No email'}
-                          </div>
-                        </div>
-                      ) : (
-                        <Badge type="gray" label="Unassigned" />
-                      )}
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                      <div className="flex justify-end space-x-2">
-                        <Link
-                          to={`/vehicles/${vehicle.id}`}
-                          className="text-blue-600 hover:text-blue-900 p-1 hover:bg-blue-50 rounded transition-colors duration-200 focus-ring"
-                          title="View Details"
-                        >
-                          <Eye className="h-4 w-4" />
-                        </Link>
-                        {user?.isAdmin && (
-                          <Link
-                            to={`/vehicles/${vehicle.id}/edit`}
-                            className="text-gray-600 hover:text-gray-900 p-1 hover:bg-gray-50 rounded transition-colors duration-200 focus-ring"
-                            title="Edit Vehicle"
-                          >
-                            <Edit className="h-4 w-4" />
-                          </Link>
-                        )}
-                      </div>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          ) : (
-            <div className="text-center py-12">
-              <div className="text-gray-500 mb-4">
-                {loading ? 'Loading vehicles...' : 
-                 (searchTerm || statusFilters.length > 0 || unassignedFilter) ? 
-                 'No vehicles match your filters' : 'No vehicles found'}
-              </div>
-              {!loading && (searchTerm || statusFilters.length > 0 || unassignedFilter) && (
-                <Button
-                  onClick={clearAllFilters}
-                  variant="secondary"
-                >
-                  Clear all filters
-                </Button>
-              )}
-            </div>
-          )}
-        </div>
-
-        {/* Pagination */}
-        {totalPages > 1 && (
-          <div className="bg-white px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6">
-            <div className="flex-1 flex justify-between sm:hidden">
-              <button
-                onClick={() => handlePageChange(currentPage - 1)}
-                disabled={!hasPreviousPage || loading}
-                className="relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                Previous
-              </button>
-              <button
-                onClick={() => handlePageChange(currentPage + 1)}
-                disabled={!hasNextPage || loading}
-                className="ml-3 relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                Next
-              </button>
-            </div>
-            <div className="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
-              <div>
-                <p className="text-sm text-gray-700">
-                  Showing <span className="font-medium">{((currentPage - 1) * itemsPerPage) + 1}</span> to{' '}
-                  <span className="font-medium">{Math.min(currentPage * itemsPerPage, totalCount)}</span> of{' '}
-                  <span className="font-medium">{totalCount}</span> results
-                </p>
-              </div>
-              <div>
-                <nav className="relative z-0 inline-flex rounded-md shadow-sm -space-x-px" aria-label="Pagination">
-                  <button
-                    onClick={() => handlePageChange(currentPage - 1)}
-                    disabled={!hasPreviousPage || loading}
-                    className="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                  >
-                    <span className="sr-only">Previous</span>
-                    <ChevronLeft className="h-5 w-5" aria-hidden="true" />
-                  </button>
-                  
-                  {getPageNumbers().map((pageNumber) => (
-                    <button
-                      key={pageNumber}
-                      onClick={() => handlePageChange(pageNumber)}
-                      disabled={loading}
-                      className={`relative inline-flex items-center px-4 py-2 border text-sm font-medium ${
-                        currentPage === pageNumber
-                          ? 'z-10 bg-blue-50 border-blue-500 text-blue-600'
-                          : 'bg-white border-gray-300 text-gray-500 hover:bg-gray-50'
-                      } disabled:opacity-50 disabled:cursor-not-allowed`}
-                    >
-                      {pageNumber}
-                    </button>
-                  ))}
-                  
-                  <button
-                    onClick={() => handlePageChange(currentPage + 1)}
-                    disabled={!hasNextPage || loading}
-                    className="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                  >
-                    <span className="sr-only">Next</span>
-                    <ChevronRight className="h-5 w-5" aria-hidden="true" />
-                  </button>
-                </nav>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
+  return <VehiclesPage />;
 }
